extract components from NDSProvider
