+++ src/NDSProvider/Reset.tsx
@@ -1,0 +1,12 @@
+
+import {createGlobalStyle} from "styled-components";
+
+const Reset = createGlobalStyle(() => {
+  return {
+    body: {
+      margin: 0,
+    },
+  };
+});
+
+export default Reset;
